Added Lazcos Wrapping
Previously the filtering would ignore pixels outside the valid range, but not update their normalization. This commit adds wrapping such that invalid pixels are wrapped around to a valid index. Ideally, the filtering would take an enum as an argument and either wrap or clamp (but with correct normalization!), possibly based on the texture import settings